jasper, jstl dependency. yml 설정

+++ src/main/java/com/example/demo/Controller/UsrArticleController.java
@@ -16,64 +16,64 @@ import com.example.demo.vo.ResultData;
 
 @Controller
 public class UsrArticleController {
-	
+
 	private ArticleService articleService;
-	
+
 	@Autowired
 	public UsrArticleController(ArticleService articleService) {
 		this.articleService = articleService;
 	}
-	
+
 	// 액션 메서드
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
 	public ResultData<Article> doAdd(HttpSession session, String title, String body) {
-		
+
 		if (session.getAttribute("loginedMemberId") == null) {
-			return ResultData.from("F-A", "로그인 후 이용하세요");
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
 		}
-		
+
 		if (Util.empty(title)) {
-			return ResultData.from("F-1", "제목을 입력해주세요.");
+			return ResultData.from("F-1", "제목을 입력해주세요");
 		}
 
 		if (Util.empty(body)) {
-			return ResultData.from("F-2", "내용을 입력해주세요.");
+			return ResultData.from("F-2", "내용을 입력해주세요");
 		}
-		
+
 		articleService.writeArticle((int) session.getAttribute("loginedMemberId"), title, body);
-		
+
 		int id = articleService.getLastInsertId();
-		
-		return ResultData.from("S-1", Util.f("%d번 게시글이 생성되었습니다", id), articleService.getArticleById(id));
+
+		return ResultData.from("S-1", Util.f("%d번 게시글이 생성되었습니다", id), "article", articleService.getArticleById(id));
 	}
 
 	@RequestMapping("/usr/article/getArticles")
 	@ResponseBody
 	public ResultData<List<Article>> getArticles() {
-		
+
 		List<Article> articles = articleService.getArticles();
-		
+
 		if (articles.size() == 0) {
-			return ResultData.from("F-1", "게시글이 존재하지 않습니다");
+			return ResultData.from("F-1", "게시물이 존재하지 않습니다");
 		}
-		
-		return ResultData.from("S-1", "게시글 리스트", articles) ;
+
+		return ResultData.from("S-1", "게시물 리스트", "articles", articles);
 	}
-	
+
 	@RequestMapping("/usr/article/getArticle")
 	@ResponseBody
 	public ResultData<Article> getArticle(int id) {
-		
+
 		Article foundArticle = articleService.getArticleById(id);
-		
+
 		if (foundArticle == null) {
 			return ResultData.from("F-1", Util.f("%d번 게시글은 존재하지 않습니다", id));
 		}
-		
-		return ResultData.from("S-1", Util.f("%d번 게시글 입니다", id), foundArticle);
+
+		return ResultData.from("S-1", Util.f("%d번 게시글 입니다", id), "article", foundArticle);
 	}
-	
+
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
 	public ResultData<Article> doModify(HttpSession session, int id, String title, String body) {
@@ -82,10 +82,14 @@ public class UsrArticleController {
 			return ResultData.from("F-A", "로그인 후 이용해주세요");
 		}
 
+		if(Util.empty(id)) {
+			return ResultData.from("F-1", "수정할 글 번호를 입력해주세요");
+		}
+		
 		Article foundArticle = articleService.getArticleById(id);
 
 		if (foundArticle == null) {
-			return ResultData.from("F-1", Util.f("%d번 게시글은 존재하지 않습니다", id));
+			return ResultData.from("F-2", Util.f("%d번 게시글은 존재하지 않습니다", id));
 		}
 
 		ResultData actorCanModifyRd = articleService.actorCanModify((int) session.getAttribute("loginedMemberId"),
@@ -101,23 +105,27 @@ public class UsrArticleController {
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
 	public ResultData doDelete(HttpSession session, int id) {
-		
+
 		if (session.getAttribute("loginedMemberId") == null) {
-			return ResultData.from("F-A", "로그인 후 이용하세요");
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
 		}
 		
+		if(Util.empty(id)) {
+			return ResultData.from("F-1", "삭제할 글 번호를 입력해주세요");
+		}
+
 		Article foundArticle = articleService.getArticleById(id);
-		
+
 		if (foundArticle == null) {
-			return ResultData.from("F-1", Util.f("%d번 게시글은 존재하지 않습니다", id));
+			return ResultData.from("F-2", Util.f("%d번 게시글은 존재하지 않습니다", id));
 		}
-		
+
 		if ((int) session.getAttribute("loginedMemberId") != foundArticle.getMemberId()) {
 			return ResultData.from("F-B", "해당 게시글에 대한 권한이 없습니다");
 		}
-		
+
 		articleService.deleteArticle(id);
-		
+
 		return ResultData.from("S-1", Util.f("%d번 게시글을 삭제했습니다", id));
 	}
 }
